+++ lib/src/interpolation/interpolation.dart
@@ -7,8 +7,7 @@ import 'package:equations/equations.dart';
 ///
 ///  - [LinearInterpolation];
 ///  - [PolynomialInterpolation];
-///  - SplineInterpolation;
-///  - NewtonInterpolation.
+///  - [NewtonInterpolation];
 ///
 /// Subclasses of [Interpolation] should only override the `compute()` method.
 abstract class Interpolation {